Feat: divided
